+++ FRITTTE/src/main/java/com/revature/frittte/util/web/dto/CCInitializer.java
@@ -15,7 +15,7 @@ public class CCInitializer {
     private String creditCardName;
     private int cvv;
     private String expDate;
-    private double limit;
+    private int limit;
     private String customerUsername;
 
 }
